test/requirements.txt
glfw
PyOpenGL
PyGLM
Pillow
